Refactor EAS utilities into focused modules
